// Web/src/swagger-gen/model/pagingResponseReviewerReportsResponse.ts
/**
 * Higgs API
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { ReviewerReportsResponse } from './reviewerReportsResponse';


export interface PagingResponseReviewerReportsResponse {
    pageNumber?: number;
    pageSize?: number;
    totalPages?: number;
    data?: Array<ReviewerReportsResponse>;
}